+++ src/main/java/soma/haeya/lms/classroom/controller/ClassroomController.java
@@ -8,11 +8,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import soma.haeya.lms._config.advice.AccountId;
 import soma.haeya.lms.classroom.model.response.MyClassroomWithCountResponse;
 import soma.haeya.lms.classroom.models.request.ClassroomCreateRequest;
 import soma.haeya.lms.classroom.models.response.ClassroomResponse;
 import soma.haeya.lms.classroom.service.ClassroomService;
-import soma.haeya.lms.common.config.advice.AccountId;
 import soma.haeya.lms.common.models.response.SuccessResponse;
 
 @RestController
